% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elja.R
\name{ELJAlogistic}
\alias{ELJAlogistic}
\title{Logistic regression tool for EnvWAS/EWAS analysis}
\usage{
ELJAlogistic(
  var,
  var_adjust = NULL,
  data,
  manplot = TRUE,
  nbvalmanplot = 100,
  Bonferroni = FALSE,
  FDR = FALSE,
  manplotsign = FALSE
)
}
\arguments{
\item{var}{A categorical and binary variable. It is generally your outcome.}

\item{var_adjust}{A vector containing the names of the fixed adjustment variables for all the models.}

\item{data}{A dataframe containing all the variables needed for the analysis.}

\item{manplot}{Generate a Manhattan plot of the results of the analysis.}

\item{nbvalmanplot}{The number of variables to include in each Manhattan plot.}

\item{Bonferroni}{Add a dashed bar to the Manhattan plot showing the Bonferroni significance level.}

\item{FDR}{Add a dashed bar to the Manhattan plot showing the False Discovery Rate (Benjamini-Hochberg method) significance threshold. NA if all p-values > FDR corrected p-values.}

\item{manplotsign}{Generates a Manhattan plot with only significant results (p<0.05).}
}
\value{
A Dataframe with results for each variable of the model.
}
\description{
A tool for Environment-Wide Association Studies (EnvWAS / EWAS) which are repeated analysis. This function is espacially for logistic regression based on the glm function with a binomial family with a logit link and allows the addition of adjustment variables.
}
\examples{
### Loading the synthetic diabetes dataset contained in the mlbench package

library(mlbench)
data(SynthDiabetes)

### Using ELJAlinear to perform EWAS analysis

ELJAlogistic(var = 'diabetes',data = SynthDiabetes,manplot = TRUE,
Bonferroni = TRUE,FDR = TRUE, nbvalmanplot = 30, manplotsign = FALSE)
results

}
\references{
Dunn OJ. Multiple Comparisons Among Means. Journal of the American Statistical Association. 1961;56(293):52‑64.
Benjamini Y, Hochberg Y. Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing. Journal of the Royal Statistical Society: Series B (Methodological). 1995;57(1):289‑300.
Leisch F, Dimitriadou E. mlbench: Machine Learning Benchmark Problems. R package version 2.1-11; 2025. Available from: https://CRAN.R-project.org/package=mlbench
}
