% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elja.R
\name{ELJAglm}
\alias{ELJAglm}
\title{Generalized Linear Models regression for EnvWAS/EWAS analysis}
\usage{
ELJAglm(
  var,
  var_adjust = NULL,
  family = binomial(link = "logit"),
  data,
  manplot = TRUE,
  nbvalmanplot = 100,
  Bonferroni = FALSE,
  FDR = FALSE,
  manplotsign = FALSE
)
}
\arguments{
\item{var}{A categorical and binary variable. It is generally your outcome.}

\item{var_adjust}{A vector containing the names of the fixed adjustment variables for all the models.}

\item{family}{The family and the link use for the glm function.}

\item{data}{A dataframe containing all the variables needed for the analysis.}

\item{manplot}{Generate a Manhattan plot of the results of the analysis.}

\item{nbvalmanplot}{The number of variables to include in each Manhattan plot.}

\item{Bonferroni}{Add a dashed bar to the Manhattan plot showing the Bonferroni significance threshold.}

\item{FDR}{Add a dashed bar to the Manhattan plot showing the False Discovery Rate (Benjamini-Hochberg method) significance threshold. NA if all p-values > FDR corrected p-values.}

\item{manplotsign}{Generates a Manhattan plot with only significant results (p<0.05).}
}
\value{
A Dataframe with results for each variable of the model.
}
\description{
A tool for Environment-Wide Association Studies (EnvWAS / EWAS) which are repeated analysis. This function is espacially for generalized linear models 'glm' and allows the addition of adjustment variables.
}
\examples{
### Creating a dataframe with random variables


exposure1 <- as.factor(sample(c("Always", "Often","Never"), size = 400, replace = TRUE))
exposure2 <- as.factor(sample(c("Always", "Often","Never"), size = 400, replace = TRUE))
exposure3 <- as.factor(sample(c("Exposed", "Not exposed"), size = 400, replace = TRUE))
exposure4 <- as.factor(sample(c("Yes", "No"), size = 400, replace = TRUE))
exposure5 <- as.factor(sample(c("Yes", "No"), size = 400, replace = TRUE))
exposure6 <- as.factor(sample(c("Yes", "No"), size = 400, replace = TRUE))
exposure7 <- as.numeric(sample(0:400, 400, rep = TRUE))
exposure8 <- as.numeric(sample(0:400, 400, rep = TRUE))
exposure9 <- as.numeric(sample(0:0.1, 400, rep = TRUE))
outcome <- as.factor(sample(c("Yes", "No"), size = 400, replace = TRUE))

data <- data.frame(exposure1,exposure2,exposure3,exposure4,exposure5,
                   exposure6,exposure7,exposure8,exposure9, outcome)


### Launch of the EnvWAS analysis

ELJAglm(var = 'outcome',data = data, var_adjust = c('exposure5'))

}
\references{
\enumerate{
\item Dunn OJ. Multiple Comparisons Among Means. Journal of the American Statistical Association. 1961;56(293):52‑64.
\item Benjamini Y, Hochberg Y. Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing. Journal of the Royal Statistical Society: Series B (Methodological). 1995;57(1):289‑300.
}
}
